refactor onboarding agree
