+++ app/constants/dimensions.js
@@ -5,9 +5,13 @@ export default {
 
   ICON_SIZE: 32,
 
+  LENGTH_250: 250,
+
   MARGIN_XSMALL: 2,
   MARGIN_SMALL: 4,
   MARGIN: 8,
+  MARGIN_LARGE: 12,
+  MARGIN_XLARGE: 16,
 
   PADDING_SMALL: 4,
   PADDING: 8
