+++ app/screens/OnboardingAgree/container.js
@@ -1,0 +1,17 @@
+import { connect } from 'react-redux'
+import { goToOnboardingRequest } from 'hg/actions/navigation'
+import component from './component'
+
+const mapDispatchToProps = dispatch => {
+  return {
+    goToOnboardingRequest: () => {
+      dispatch(goToOnboardingRequest())
+    }
+  }
+}
+
+const mapStateToProps = () => {
+  return {}
+}
+
+export default connect(mapStateToProps,mapDispatchToProps)(component)
