+++ app/screens/OnboardingAgree/index.js
@@ -1,4 +1,1 @@
-import OnboardingAgree from './OnboardingAgree'
-import { connect } from './presenter'
-
-export default connect(OnboardingAgree)
+export { default } from './container'
