+++ app/screens/Splash/container.js
@@ -9,4 +9,4 @@ const mapStateToProps = (state) => {
   return {}
 }
 
-export default connect(mapStateToProps,mapDispatchToProps)(component)
+export default connect(mapStateToProps, mapDispatchToProps)(component)
